Add solution for day 11. Let `Coor` class use Long instead of Int.

+++ src/main/kotlin/com/cormontia/adventOfCode/year2023/utils/BuildMap.kt
@@ -11,7 +11,7 @@ fun buildGridMap(strings: List<String>): MutableMap<Coor, Char> {
         val row = strings[rowIdx]
         for (colIdx in row.indices) {
             val char = row[colIdx]
-            map[Coor(rowIdx, colIdx)] = char
+            map[Coor(rowIdx.toLong(), colIdx.toLong())] = char
         }
     }
     return map
